LRDOCS-6340 Create util file for popular methods
