+++ code/liferay-doc-utils/src/com/liferay/documentation/util/ConvertHeadersTask.java
@@ -4,14 +4,10 @@ import java.io.BufferedWriter;
 import java.io.File;
 import java.io.FileReader;
 import java.io.FileWriter;
-import java.io.FilenameFilter;
 import java.io.IOException;
 import java.io.LineNumberReader;
 import java.util.ArrayList;
-import java.util.LinkedList;
 import java.util.List;
-import java.util.Queue;
-import java.util.regex.Pattern;
 
 import org.apache.commons.io.FileUtils;
 import org.apache.tools.ant.BuildException;
@@ -23,8 +19,8 @@ public class ConvertHeadersTask extends Task {
 
 		String docDir = _docdir;
 
-		List<String> ceFileList = getFileList(docDir, "");
-		List<String> dxpFileList = getFileList(docDir, "-dxp");
+		List<String> ceFileList = DocsUtil.getMarkdownFileList(docDir, "");
+		List<String> dxpFileList = DocsUtil.getMarkdownFileList(docDir, "-dxp");
 		
 		List<String> fileList = new ArrayList<String>();
 		fileList.addAll(ceFileList);
@@ -95,68 +91,6 @@ public class ConvertHeadersTask extends Task {
 			}
 		}
 		
-	}
-
-	private static List<String> getFileList(String docDir, String dirType)
-			throws BuildException {
-		
-		File articlesDir = new File("../" + docDir + "/articles" + dirType);
-		File docSetDir = new File("../" + docDir);
-
-		if (!articlesDir.exists() && !dirType.contains("dxp")) {
-			throw new BuildException("FAILURE - bad articles directory " + articlesDir);
-		}
-		else if (!articlesDir.exists() && dirType.contains("dxp")) {
-			return new ArrayList<String>();
-		}
-
-		List<File> docSetDirFolders = new ArrayList<File>();
-		Queue<File> q = new LinkedList<File>();
-
-		File articlesDirContents[] = articlesDir.listFiles();
-		for (File f : articlesDirContents) {
-			if (f.isDirectory()) {
-				q.add(f);
-			}
-		}
-	
-		while (!q.isEmpty()) {
-			File f = q.remove();
-			docSetDirFolders.add(f);
-			File[] files = f.listFiles();
-		
-			for (File file : files) {
-				if (file.isDirectory()) {
-					q.add(file);
-				}
-			}
-		}
-
-		docSetDirFolders.add(articlesDir);
-		docSetDirFolders.add(docSetDir);
-	
-		File docSetDirFoldersArray[] = docSetDirFolders.toArray(new File[docSetDirFolders.size()]);
-	
-		List<String> fileList = new ArrayList<String>();
-
-		for (int i = 0; i < docSetDirFoldersArray.length; i++) {
-			File files[] = docSetDirFoldersArray[i].listFiles(new FilenameFilter() {
-				String filePatternArg =
-						"([^\\\\\\[\\]\\|:;%<>]+).markdown";
-				Pattern fileNamePattern = Pattern.compile(filePatternArg);
-
-				public boolean accept(File file, String name) {
-					return (fileNamePattern.matcher(name).matches());
-				}
-			});
-		
-			for (int j = 0; j < files.length; j++) {
-				fileList.add(files[j].getPath());
-			}
-		
-		}
-		
-		return fileList;
 	}
 
 	public void setDocdir(String docdir) {
